requirements.txt
Flask==3.1.0
line_bot_sdk==3.14.2
